+++ src/fphysics/collision/detect-circle-circle-collision.ts
@@ -36,5 +36,6 @@ export function detectCircleCircleCollision(
     normal,
     depth,
     contactPoints: [contactPoint],
+    featureIds: [0],
   };
 }