+++ src/fphysics/index.ts
@@ -2,5 +2,6 @@ export * from './apply-impluse.js';
 export * from './colliders/index.js';
 export * from './collision/index.js';
 export * from './components/index.js';
+export * from './solve-soft-constraint.js';
 export * from './systems/index.js';
 export * from './types/index.js';